backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pinecone-client==3.0.0
sentence-transformers==2.7.0
python-dotenv==1.0.0
requests==2.31.0
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
numpy<2.0.0
huggingface-hub==0.20.0
